вынес перечесления в файл constants.js
