+++ src/util.js
@@ -1,26 +1,6 @@
-const SITE_BODY = document.querySelector(`body`);
+import {IterationCount, InsertPlace} from "./constants.js";
 
-const IterationCount = Object.freeze({
-  DEFAULT: 1,
-  CARD: 5,
-  EXTRA: 2,
-  COMMENT: 4
-});
-
-const InsertPlace = Object.freeze({
-  BEFOREEND: `beforeend`,
-  AFTERBEGIN: `afterbegin`,
-  AFTEREND: `afterend`
-});
-
-const SiteElements = Object.freeze({
-  BODY: SITE_BODY,
-  HEADER: SITE_BODY.querySelector(`.header`),
-  MAIN: SITE_BODY.querySelector(`main`),
-  FOOTER: SITE_BODY.querySelector(`.footer`)
-});
-
-const render = ({
+export const render = ({
   container,
   template,
   place = InsertPlace.BEFOREEND,
@@ -30,5 +10,3 @@ const render = ({
     container.insertAdjacentHTML(place, template);
   }
 };
-
-export {IterationCount, InsertPlace, SiteElements, render};
