+++ webpack.config.js
@@ -5,7 +5,7 @@ module.exports = {
   entry: './src/main.js',
   output: {
     filename: 'bundle.js',
-    path: path.join(__dirname, 'public/js'),
+    path: path.join(__dirname, 'public'),
   },
   devtool: 'source-map',
   devServer: {
